refactor(#792): split audio_signal_integrity_tests into base + EQ suite (sweep)
audio_signal_integrity_tests.rs 778 -> 508, + audio_signal_integrity_eq_tests.rs 279
(both under cap; off debt). The EQ signal-integrity tests + eq_with_one_band_boosted
move to the sibling; the shared runtime/scan fixtures (build_runtime, drive_capture_steady,
chain_with_blocks, core_block, scan_*, input_mono, output, SineGen, SR, BUFFER_FRAMES)
become pub(super) and the new file imports them via super::audio_signal_integrity.
engine tests (36 groups) pass, zero warnings.
